backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
pdfplumber==0.10.3
qdrant-client==1.7.0
sentence-transformers==2.2.2
groq==0.4.1
python-dotenv==1.0.0
python-dotenv==1.1.1